Restructure `docs` crate
