+++ crates/docs/src/lib.rs
@@ -1,2 +1,7 @@
 #[cfg(feature = "testing")]
+pub mod snippet;
+#[cfg(feature = "testing")]
 pub mod validation;
+
+#[cfg(feature = "testing")]
+pub mod utils;
